backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
duckduckgo-search
google-generativeai
pydantic
